requirements.txt
aiohttp==3.8.3
aioredis==2.0.1
fastapi==0.92.0
loguru==0.6.0
openai==0.26.5
pydantic==1.10.2
PyJWT==2.6.0
python-dotenv==1.0.0
revChatGPT==2.3.11
starlette==0.25.0
uvicorn==0.20.0
wechatpy==1.8.18
